third_party/WebKit/LayoutTests/fast/dom/ping-attribute-dom-binding-expected.txt

Tests the DOM bindings for the ping attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS anchor.ping is ""
PASS anchor.ping is "p1"
PASS anchor.getAttribute("ping") is "p2"
PASS anchor.getAttribute("ping") is "null"
PASS anchor.ping is "null"
PASS area.ping is ""
PASS area.ping is "p1"
PASS area.getAttribute("ping") is "p2"
PASS area.getAttribute("ping") is "null"
PASS area.ping is "null"
PASS successfullyParsed is true

TEST COMPLETE
